backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
anthropic==0.42.0
google-cloud-speech==2.28.0
